requirements.txt
spacy>=3.7.2
nltk>=3.8.1
pdf2image>=1.16.3
pytesseract>=0.3.10
python-docx>=1.0.1
beautifulsoup4>=4.12.2
transformers>=4.36.2
torch>=2.1.2
numpy>=1.24.3
pandas>=2.1.3
scikit-learn>=1.3.2
python-magic>=0.4.27
PyPDF2>=3.0.0 